Make growth chart steeper
