+++ components/GrowthSvg.js
@@ -1,0 +1,25 @@
+// SVG exported from https://www.magicpattern.design/tools/svg-chart-generator,
+// data generated with `(a := [1.1**i + random.randrange(0, int(i**3.3)+1) for i in range(200)])+[a[-1]]*5`
+// (parameters are super arbitrary)
+const GrowthSvg = () => (
+  <svg
+    width="100%"
+    height="15rem" // Set height less than container, since I can't figure out how to fit it to the container height
+    preserveAspectRatio="none"
+    viewBox="0 0 990 500"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M 0,499.9999972448506 C 0,499.9999972448506 4.878048780487805,499.99999696933565 4.878048780487805,499.99999696933565 C 4.878048780487805,499.99999696933565 9.75609756097561,499.9999801353729 9.75609756097561,499.9999801353729 C 9.75609756097561,499.9999801353729 14.634146341463413,499.9999081681156 14.634146341463413,499.9999081681156 C 14.634146341463413,499.9999081681156 19.51219512195122,499.9999215771522 19.51219512195122,499.9999215771522 C 19.51219512195122,499.9999215771522 24.390243902439025,499.9995685146485 24.390243902439025,499.9995685146485 C 24.390243902439025,499.9995685146485 29.268292682926827,499.99944684435565 29.268292682926827,499.99944684435565 C 29.268292682926827,499.99944684435565 34.146341463414636,499.99970258515765 34.146341463414636,499.99970258515765 C 34.146341463414636,499.99970258515765 39.02439024390244,499.9997681718424 39.02439024390244,499.9997681718424 C 39.02439024390244,499.9997681718424 43.90243902439024,499.9983872514059 43.90243902439024,499.9983872514059 C 43.90243902439024,499.9983872514059 48.78048780487805,499.9972266838618 48.78048780487805,499.9972266838618 C 48.78048780487805,499.9972266838618 53.65853658536585,499.9997000934016 53.65853658536585,499.9997000934016 C 53.65853658536585,499.9997000934016 58.536585365853654,499.99311174512746 58.536585365853654,499.99311174512746 C 58.536585365853654,499.99311174512746 63.41463414634146,499.9992658841868 63.41463414634146,499.9992658841868 C 63.41463414634146,499.9992658841868 68.29268292682927,499.9856131677938 68.29268292682927,499.9856131677938 C 68.29268292682927,499.9856131677938 73.17073170731707,499.99983420269126 73.17073170731707,499.99983420269126 C 73.17073170731707,499.99983420269126 78.04878048780488,499.9839220640543 78.04878048780488,499.9839220640543 C 78.04878048780488,499.9839220640543 82.92682926829268,499.99167929876035 82.92682926829268,499.99167929876035 C 82.92682926829268,499.99167929876035 87.80487804878048,499.9983150610652 87.80487804878048,499.9983150610652 C 87.80487804878048,499.9983150610652 92.6829268292683,499.9900287950012 92.6829268292683,499.9900287950012 C 92.6829268292683,499.9900287950012 97.5609756097561,499.94986805242934 97.5609756097561,499.94986805242934 C 97.5609756097561,499.94986805242934 102.4390243902439,499.9826745178705 102.4390243902439,499.9826745178705 C 102.4390243902439,499.9826745178705 107.3170731707317,499.98788522164205 107.3170731707317,499.98788522164205 C 107.3170731707317,499.98788522164205 112.1951219512195,499.9368025091346 112.1951219512195,499.9368025091346 C 112.1951219512195,499.9368025091346 117.07317073170731,499.92325848282553 117.07317073170731,499.92325848282553 C 117.07317073170731,499.92325848282553 121.95121951219512,499.9874590153744 121.95121951219512,499.9874590153744 C 121.95121951219512,499.9874590153744 126.82926829268293,499.9003382064549 126.82926829268293,499.9003382064549 C 126.82926829268293,499.9003382064549 131.70731707317074,499.85785051918424 131.70731707317074,499.85785051918424 C 131.70731707317074,499.85785051918424 136.58536585365854,499.9080457291194 136.58536585365854,499.9080457291194 C 136.58536585365854,499.9080457291194 141.46341463414635,499.84501495841357 141.46341463414635,499.84501495841357 C 141.46341463414635,499.84501495841357 146.34146341463415,499.9554287101044 146.34146341463415,499.9554287101044 C 146.34146341463415,499.9554287101044 151.21951219512195,499.8180852156077 151.21951219512195,499.8180852156077 C 151.21951219512195,499.8180852156077 156.09756097560975,499.9153394559902 156.09756097560975,499.9153394559902 C 156.09756097560975,499.9153394559902 160.97560975609755,499.94833481825464 160.97560975609755,499.94833481825464 C 160.97560975609755,499.94833481825464 165.85365853658536,499.71747720691536 165.85365853658536,499.71747720691536 C 165.85365853658536,499.71747720691536 170.73170731707316,499.7170734266381 170.73170731707316,499.7170734266381 C 170.73170731707316,499.7170734266381 175.60975609756096,499.6856680015181 175.60975609756096,499.6856680015181 C 175.60975609756096,499.6856680015181 180.48780487804876,499.685204884963 180.48780487804876,499.685204884963 C 180.48780487804876,499.685204884963 185.3658536585366,499.924554630154 185.3658536585366,499.924554630154 C 185.3658536585366,499.924554630154 190.2439024390244,499.75343577166694 190.2439024390244,499.75343577166694 C 190.2439024390244,499.75343577166694 195.1219512195122,499.76658027401595 195.1219512195122,499.76658027401595 C 195.1219512195122,499.76658027401595 200,499.4926288106007 200,499.4926288106007 C 200,499.4926288106007 204.8780487804878,499.70323248936205 204.8780487804878,499.70323248936205 C 204.8780487804878,499.70323248936205 209.7560975609756,499.59300867028867 209.7560975609756,499.59300867028867 C 209.7560975609756,499.59300867028867 214.6341463414634,499.89884671762854 214.6341463414634,499.89884671762854 C 214.6341463414634,499.89884671762854 219.5121951219512,499.56831798459183 219.5121951219512,499.56831798459183 C 219.5121951219512,499.56831798459183 224.390243902439,499.3470924677361 224.390243902439,499.3470924677361 C 224.390243902439,499.3470924677361 229.2682926829268,499.16418631554126 229.2682926829268,499.16418631554126 C 229.2682926829268,499.16418631554126 234.14634146341461,499.2104788322566 234.14634146341461,499.2104788322566 C 234.14634146341461,499.2104788322566 239.02439024390245,499.8040765907036 239.02439024390245,499.8040765907036 C 239.02439024390245,499.8040765907036 243.90243902439025,499.17517607371195 243.90243902439025,499.17517607371195 C 243.90243902439025,499.17517607371195 248.78048780487805,499.66901528001154 248.78048780487805,499.66901528001154 C 248.78048780487805,499.66901528001154 253.65853658536585,499.375757418371 253.65853658536585,499.375757418371 C 253.65853658536585,499.375757418371 258.5365853658536,498.74430416620714 258.5365853658536,498.74430416620714 C 258.5365853658536,498.74430416620714 263.4146341463415,499.2039467733315 263.4146341463415,499.2039467733315 C 263.4146341463415,499.2039467733315 268.2926829268293,499.2822338275746 268.2926829268293,499.2822338275746 C 268.2926829268293,499.2822338275746 273.1707317073171,499.3080112645127 273.1707317073171,499.3080112645127 C 273.1707317073171,499.3080112645127 278.0487804878049,498.8614792528015 278.0487804878049,498.8614792528015 C 278.0487804878049,498.8614792528015 282.9268292682927,499.31890602707665 282.9268292682927,499.31890602707665 C 282.9268292682927,499.31890602707665 287.8048780487805,499.0654180559883 287.8048780487805,499.0654180559883 C 287.8048780487805,499.0654180559883 292.6829268292683,498.73923680062745 292.6829268292683,498.73923680062745 C 292.6829268292683,498.73923680062745 297.5609756097561,499.99770239880837 297.5609756097561,499.99770239880837 C 297.5609756097561,499.99770239880837 302.4390243902439,499.18370868340645 302.4390243902439,499.18370868340645 C 302.4390243902439,499.18370868340645 307.3170731707317,499.9013511457095 307.3170731707317,499.9013511457095 C 307.3170731707317,499.9013511457095 312.1951219512195,499.0858199645777 312.1951219512195,499.0858199645777 C 312.1951219512195,499.0858199645777 317.0731707317073,499.06368074354396 317.0731707317073,499.06368074354396 C 317.0731707317073,499.06368074354396 321.9512195121951,498.99169685322136 321.9512195121951,498.99169685322136 C 321.9512195121951,498.99169685322136 326.8292682926829,499.7900511455965 326.8292682926829,499.7900511455965 C 326.8292682926829,499.7900511455965 331.7073170731707,498.0341769841697 331.7073170731707,498.0341769841697 C 331.7073170731707,498.0341769841697 336.5853658536585,498.4577777088165 336.5853658536585,498.4577777088165 C 336.5853658536585,498.4577777088165 341.4634146341463,498.4776869826434 341.4634146341463,498.4776869826434 C 341.4634146341463,498.4776869826434 346.3414634146341,498.14521215454977 346.3414634146341,498.14521215454977 C 346.3414634146341,498.14521215454977 351.2195121951219,497.7417897569322 351.2195121951219,497.7417897569322 C 351.2195121951219,497.7417897569322 356.0975609756097,496.3743028764938 356.0975609756097,496.3743028764938 C 356.0975609756097,496.3743028764938 360.9756097560975,499.6456567416484 360.9756097560975,499.6456567416484 C 360.9756097560975,499.6456567416484 365.8536585365854,496.95618287313204 365.8536585365854,496.95618287313204 C 365.8536585365854,496.95618287313204 370.7317073170732,497.00982786475356 370.7317073170732,497.00982786475356 C 370.7317073170732,497.00982786475356 375.609756097561,497.92776399517953 375.609756097561,497.92776399517953 C 375.609756097561,497.92776399517953 380.4878048780488,495.5160828237326 380.4878048780488,495.5160828237326 C 380.4878048780488,495.5160828237326 385.3658536585366,498.6601097355852 385.3658536585366,498.6601097355852 C 385.3658536585366,498.6601097355852 390.2439024390244,499.748464778396 390.2439024390244,499.748464778396 C 390.2439024390244,499.748464778396 395.1219512195122,495.2962181896429 395.1219512195122,495.2962181896429 C 395.1219512195122,495.2962181896429 400,498.9607285429721 400,498.9607285429721 C 400,498.9607285429721 404.8780487804878,498.7220117745617 404.8780487804878,498.7220117745617 C 404.8780487804878,498.7220117745617 409.7560975609756,498.09977604612914 409.7560975609756,498.09977604612914 C 409.7560975609756,498.09977604612914 414.6341463414634,499.43071458691867 414.6341463414634,499.43071458691867 C 414.6341463414634,499.43071458691867 419.5121951219512,497.7806696754445 419.5121951219512,497.7806696754445 C 419.5121951219512,497.7806696754445 424.390243902439,493.28987841837574 424.390243902439,493.28987841837574 C 424.390243902439,493.28987841837574 429.2682926829268,494.50419349945076 429.2682926829268,494.50419349945076 C 429.2682926829268,494.50419349945076 434.1463414634146,492.87932735861165 434.1463414634146,492.87932735861165 C 434.1463414634146,492.87932735861165 439.0243902439024,498.6915188778363 439.0243902439024,498.6915188778363 C 439.0243902439024,498.6915188778363 443.9024390243902,498.10194913682335 443.9024390243902,498.10194913682335 C 443.9024390243902,498.10194913682335 448.780487804878,498.1148392892645 448.780487804878,498.1148392892645 C 448.780487804878,498.1148392892645 453.6585365853658,491.37298906777175 453.6585365853658,491.37298906777175 C 453.6585365853658,491.37298906777175 458.5365853658536,498.38378584480256 458.5365853658536,498.38378584480256 C 458.5365853658536,498.38378584480256 463.4146341463414,490.9599176664429 463.4146341463414,490.9599176664429 C 463.4146341463414,490.9599176664429 468.29268292682923,497.0303860766468 468.29268292682923,497.0303860766468 C 468.29268292682923,497.0303860766468 473.1707317073171,495.8414585686672 473.1707317073171,495.8414585686672 C 473.1707317073171,495.8414585686672 478.0487804878049,497.3599884772825 478.0487804878049,497.3599884772825 C 478.0487804878049,497.3599884772825 482.9268292682927,496.36298561490514 482.9268292682927,496.36298561490514 C 482.9268292682927,496.36298561490514 487.8048780487805,494.44655664323653 487.8048780487805,494.44655664323653 C 487.8048780487805,494.44655664323653 492.6829268292683,488.88054581332267 492.6829268292683,488.88054581332267 C 492.6829268292683,488.88054581332267 497.5609756097561,491.2695334620059 497.5609756097561,491.2695334620059 C 497.5609756097561,491.2695334620059 502.4390243902439,493.425238239918 502.4390243902439,493.425238239918 C 502.4390243902439,493.425238239918 507.3170731707317,495.87710023778203 507.3170731707317,495.87710023778203 C 507.3170731707317,495.87710023778203 512.1951219512194,487.73846987963793 512.1951219512194,487.73846987963793 C 512.1951219512194,487.73846987963793 517.0731707317073,499.13728504267885 517.0731707317073,499.13728504267885 C 517.0731707317073,499.13728504267885 521.9512195121952,488.4499494621698 521.9512195121952,488.4499494621698 C 521.9512195121952,488.4499494621698 526.829268292683,496.2089558554846 526.829268292683,496.2089558554846 C 526.829268292683,496.2089558554846 531.7073170731708,498.97733548479164 531.7073170731708,498.97733548479164 C 531.7073170731708,498.97733548479164 536.5853658536586,487.618079369193 536.5853658536586,487.618079369193 C 536.5853658536586,487.618079369193 541.4634146341464,487.61979490171296 541.4634146341464,487.61979490171296 C 541.4634146341464,487.61979490171296 546.3414634146342,489.09403740462096 546.3414634146342,489.09403740462096 C 546.3414634146342,489.09403740462096 551.219512195122,492.14060460630145 551.219512195122,492.14060460630145 C 551.219512195122,492.14060460630145 556.0975609756098,489.57447377748844 556.0975609756098,489.57447377748844 C 556.0975609756098,489.57447377748844 560.9756097560976,497.03784510907695 560.9756097560976,497.03784510907695 C 560.9756097560976,497.03784510907695 565.8536585365854,487.0450353453407 565.8536585365854,487.0450353453407 C 565.8536585365854,487.0450353453407 570.7317073170732,495.6478015490989 570.7317073170732,495.6478015490989 C 570.7317073170732,495.6478015490989 575.609756097561,489.2071952464976 575.609756097561,489.2071952464976 C 575.609756097561,489.2071952464976 580.4878048780488,494.998133806539 580.4878048780488,494.998133806539 C 580.4878048780488,494.998133806539 585.3658536585366,491.11132306402413 585.3658536585366,491.11132306402413 C 585.3658536585366,491.11132306402413 590.2439024390244,493.9984113073628 590.2439024390244,493.9984113073628 C 590.2439024390244,493.9984113073628 595.1219512195122,493.86074476435454 595.1219512195122,493.86074476435454 C 595.1219512195122,493.86074476435454 600,481.10169807799264 600,481.10169807799264 C 600,481.10169807799264 604.8780487804878,477.69275289970835 604.8780487804878,477.69275289970835 C 604.8780487804878,477.69275289970835 609.7560975609756,483.36032240446684 609.7560975609756,483.36032240446684 C 609.7560975609756,483.36032240446684 614.6341463414634,498.52155440595664 614.6341463414634,498.52155440595664 C 614.6341463414634,498.52155440595664 619.5121951219512,493.0089193171211 619.5121951219512,493.0089193171211 C 619.5121951219512,493.0089193171211 624.390243902439,493.3279177466967 624.390243902439,493.3279177466967 C 624.390243902439,493.3279177466967 629.2682926829268,481.5084948825125 629.2682926829268,481.5084948825125 C 629.2682926829268,481.5084948825125 634.1463414634146,475.62608200138186 634.1463414634146,475.62608200138186 C 634.1463414634146,475.62608200138186 639.0243902439024,479.84152299056126 639.0243902439024,479.84152299056126 C 639.0243902439024,479.84152299056126 643.9024390243902,488.38952557154755 643.9024390243902,488.38952557154755 C 643.9024390243902,488.38952557154755 648.780487804878,473.03388205857647 648.780487804878,473.03388205857647 C 648.780487804878,473.03388205857647 653.6585365853658,475.9821958431486 653.6585365853658,475.9821958431486 C 653.6585365853658,475.9821958431486 658.5365853658536,496.34733372524744 658.5365853658536,496.34733372524744 C 658.5365853658536,496.34733372524744 663.4146341463414,480.12001934929145 663.4146341463414,480.12001934929145 C 663.4146341463414,480.12001934929145 668.2926829268292,481.2182065601716 668.2926829268292,481.2182065601716 C 668.2926829268292,481.2182065601716 673.170731707317,469.03509705770347 673.170731707317,469.03509705770347 C 673.170731707317,469.03509705770347 678.0487804878048,474.14251179191905 678.0487804878048,474.14251179191905 C 678.0487804878048,474.14251179191905 682.9268292682926,477.9257423612485 682.9268292682926,477.9257423612485 C 682.9268292682926,477.9257423612485 687.8048780487804,497.7899561395063 687.8048780487804,497.7899561395063 C 687.8048780487804,497.7899561395063 692.6829268292682,487.9301277267934 692.6829268292682,487.9301277267934 C 692.6829268292682,487.9301277267934 697.560975609756,464.5619018460006 697.560975609756,464.5619018460006 C 697.560975609756,464.5619018460006 702.4390243902438,470.3621621853033 702.4390243902438,470.3621621853033 C 702.4390243902438,470.3621621853033 707.3170731707316,489.7679000145479 707.3170731707316,489.7679000145479 C 707.3170731707316,489.7679000145479 712.1951219512194,465.6254005090687 712.1951219512194,465.6254005090687 C 712.1951219512194,465.6254005090687 717.0731707317073,460.1203494433305 717.0731707317073,460.1203494433305 C 717.0731707317073,460.1203494433305 721.951219512195,456.92228351442196 721.951219512195,456.92228351442196 C 721.951219512195,456.92228351442196 726.8292682926829,464.0348513067689 726.8292682926829,464.0348513067689 C 726.8292682926829,464.0348513067689 731.7073170731708,490.9784951987187 731.7073170731708,490.9784951987187 C 731.7073170731708,490.9784951987187 736.5853658536586,456.5709558724289 736.5853658536586,456.5709558724289 C 736.5853658536586,456.5709558724289 741.4634146341464,457.3796143652458 741.4634146341464,457.3796143652458 C 741.4634146341464,457.3796143652458 746.3414634146342,456.80397037291414 746.3414634146342,456.80397037291414 C 746.3414634146342,456.80397037291414 751.219512195122,467.3187053120997 751.219512195122,467.3187053120997 C 751.219512195122,467.3187053120997 756.0975609756098,477.64488106582087 756.0975609756098,477.64488106582087 C 756.0975609756098,477.64488106582087 760.9756097560976,466.9385989962091 760.9756097560976,466.9385989962091 C 760.9756097560976,466.9385989962091 765.8536585365854,451.84278120927047 765.8536585365854,451.84278120927047 C 765.8536585365854,451.84278120927047 770.7317073170732,461.4440788863084 770.7317073170732,461.4440788863084 C 770.7317073170732,461.4440788863084 775.609756097561,464.30595399115776 775.609756097561,464.30595399115776 C 775.609756097561,464.30595399115776 780.4878048780488,443.26627107308843 780.4878048780488,443.26627107308843 C 780.4878048780488,443.26627107308843 785.3658536585366,434.5579482135822 785.3658536585366,434.5579482135822 C 785.3658536585366,434.5579482135822 790.2439024390244,467.5074829744035 790.2439024390244,467.5074829744035 C 790.2439024390244,467.5074829744035 795.1219512195122,450.19865598145964 795.1219512195122,450.19865598145964 C 795.1219512195122,450.19865598145964 800,433.2313079470399 800,433.2313079470399 C 800,433.2313079470399 804.8780487804878,448.598255407735 804.8780487804878,448.598255407735 C 804.8780487804878,448.598255407735 809.7560975609756,421.82728735039757 809.7560975609756,421.82728735039757 C 809.7560975609756,421.82728735039757 814.6341463414634,444.10138084772024 814.6341463414634,444.10138084772024 C 814.6341463414634,444.10138084772024 819.5121951219512,451.97870652129455 819.5121951219512,451.97870652129455 C 819.5121951219512,451.97870652129455 824.390243902439,449.9972699170539 824.390243902439,449.9972699170539 C 824.390243902439,449.9972699170539 829.2682926829268,410.09477046376986 829.2682926829268,410.09477046376986 C 829.2682926829268,410.09477046376986 834.1463414634146,429.48871429367523 834.1463414634146,429.48871429367523 C 834.1463414634146,429.48871429367523 839.0243902439024,432.22514345900515 839.0243902439024,432.22514345900515 C 839.0243902439024,432.22514345900515 843.9024390243902,428.0562117597184 843.9024390243902,428.0562117597184 C 843.9024390243902,428.0562117597184 848.780487804878,389.74229298208843 848.780487804878,389.74229298208843 C 848.780487804878,389.74229298208843 853.6585365853658,437.50048009116165 853.6585365853658,437.50048009116165 C 853.6585365853658,437.50048009116165 858.5365853658536,403.85475631933116 858.5365853658536,403.85475631933116 C 858.5365853658536,403.85475631933116 863.4146341463414,395.551797676389 863.4146341463414,395.551797676389 C 863.4146341463414,395.551797676389 868.2926829268292,369.7432273922798 868.2926829268292,369.7432273922798 C 868.2926829268292,369.7432273922798 873.170731707317,384.2032877341777 873.170731707317,384.2032877341777 C 873.170731707317,384.2032877341777 878.0487804878048,403.50423376307435 878.0487804878048,403.50423376307435 C 878.0487804878048,403.50423376307435 882.9268292682926,361.0049249699402 882.9268292682926,361.0049249699402 C 882.9268292682926,361.0049249699402 887.8048780487804,350.15668673463193 887.8048780487804,350.15668673463193 C 887.8048780487804,350.15668673463193 892.6829268292682,358.0062111624337 892.6829268292682,358.0062111624337 C 892.6829268292682,358.0062111624337 897.560975609756,333.0138717631859 897.560975609756,333.0138717631859 C 897.560975609756,333.0138717631859 902.4390243902438,350.70142634168656 902.4390243902438,350.70142634168656 C 902.4390243902438,350.70142634168656 907.3170731707316,348.6795149633591 907.3170731707316,348.6795149633591 C 907.3170731707316,348.6795149633591 912.1951219512194,302.4633869318664 912.1951219512194,302.4633869318664 C 912.1951219512194,302.4633869318664 917.0731707317073,258.9045868612644 917.0731707317073,258.9045868612644 C 917.0731707317073,258.9045868612644 921.951219512195,295.6194339366724 921.951219512195,295.6194339366724 C 921.951219512195,295.6194339366724 926.8292682926829,248.99346745137777 926.8292682926829,248.99346745137777 C 926.8292682926829,248.99346745137777 931.7073170731707,204.73863099983237 931.7073170731707,204.73863099983237 C 931.7073170731707,204.73863099983237 936.5853658536585,170.72247507418405 936.5853658536585,170.72247507418405 C 936.5853658536585,170.72247507418405 941.4634146341464,202.11801321761766 941.4634146341464,202.11801321761766 C 941.4634146341464,202.11801321761766 946.3414634146342,118.39287368567165 946.3414634146342,118.39287368567165 C 946.3414634146342,118.39287368567165 951.219512195122,171.569985184289 951.219512195122,171.569985184289 C 951.219512195122,171.569985184289 956.0975609756098,121.00619811075836 956.0975609756098,121.00619811075836 C 956.0975609756098,121.00619811075836 960.9756097560976,98.91212364408796 960.9756097560976,98.91212364408796 C 960.9756097560976,98.91212364408796 965.8536585365854,19.464634386687692 965.8536585365854,19.464634386687692 C 965.8536585365854,19.464634386687692 970.7317073170732,0 970.7317073170732,0 C 970.7317073170732,0 975.609756097561,0 975.609756097561,0 C 975.609756097561,0 980.4878048780488,0 980.4878048780488,0 C 980.4878048780488,0 985.3658536585366,0 985.3658536585366,0 C 985.3658536585366,0 990.2439024390244,0 990.2439024390244,0 C 990.2439024390244,0 995.1219512195122,0 995.1219512195122,0,L 1000 500,L 0 500Z"
+      fill="#fbbf241a"
+    />
+    <path
+      d="M 0,499.9999972448506 C 0,499.9999972448506 4.878048780487805,499.99999696933565 4.878048780487805,499.99999696933565 C 4.878048780487805,499.99999696933565 9.75609756097561,499.9999801353729 9.75609756097561,499.9999801353729 C 9.75609756097561,499.9999801353729 14.634146341463413,499.9999081681156 14.634146341463413,499.9999081681156 C 14.634146341463413,499.9999081681156 19.51219512195122,499.9999215771522 19.51219512195122,499.9999215771522 C 19.51219512195122,499.9999215771522 24.390243902439025,499.9995685146485 24.390243902439025,499.9995685146485 C 24.390243902439025,499.9995685146485 29.268292682926827,499.99944684435565 29.268292682926827,499.99944684435565 C 29.268292682926827,499.99944684435565 34.146341463414636,499.99970258515765 34.146341463414636,499.99970258515765 C 34.146341463414636,499.99970258515765 39.02439024390244,499.9997681718424 39.02439024390244,499.9997681718424 C 39.02439024390244,499.9997681718424 43.90243902439024,499.9983872514059 43.90243902439024,499.9983872514059 C 43.90243902439024,499.9983872514059 48.78048780487805,499.9972266838618 48.78048780487805,499.9972266838618 C 48.78048780487805,499.9972266838618 53.65853658536585,499.9997000934016 53.65853658536585,499.9997000934016 C 53.65853658536585,499.9997000934016 58.536585365853654,499.99311174512746 58.536585365853654,499.99311174512746 C 58.536585365853654,499.99311174512746 63.41463414634146,499.9992658841868 63.41463414634146,499.9992658841868 C 63.41463414634146,499.9992658841868 68.29268292682927,499.9856131677938 68.29268292682927,499.9856131677938 C 68.29268292682927,499.9856131677938 73.17073170731707,499.99983420269126 73.17073170731707,499.99983420269126 C 73.17073170731707,499.99983420269126 78.04878048780488,499.9839220640543 78.04878048780488,499.9839220640543 C 78.04878048780488,499.9839220640543 82.92682926829268,499.99167929876035 82.92682926829268,499.99167929876035 C 82.92682926829268,499.99167929876035 87.80487804878048,499.9983150610652 87.80487804878048,499.9983150610652 C 87.80487804878048,499.9983150610652 92.6829268292683,499.9900287950012 92.6829268292683,499.9900287950012 C 92.6829268292683,499.9900287950012 97.5609756097561,499.94986805242934 97.5609756097561,499.94986805242934 C 97.5609756097561,499.94986805242934 102.4390243902439,499.9826745178705 102.4390243902439,499.9826745178705 C 102.4390243902439,499.9826745178705 107.3170731707317,499.98788522164205 107.3170731707317,499.98788522164205 C 107.3170731707317,499.98788522164205 112.1951219512195,499.9368025091346 112.1951219512195,499.9368025091346 C 112.1951219512195,499.9368025091346 117.07317073170731,499.92325848282553 117.07317073170731,499.92325848282553 C 117.07317073170731,499.92325848282553 121.95121951219512,499.9874590153744 121.95121951219512,499.9874590153744 C 121.95121951219512,499.9874590153744 126.82926829268293,499.9003382064549 126.82926829268293,499.9003382064549 C 126.82926829268293,499.9003382064549 131.70731707317074,499.85785051918424 131.70731707317074,499.85785051918424 C 131.70731707317074,499.85785051918424 136.58536585365854,499.9080457291194 136.58536585365854,499.9080457291194 C 136.58536585365854,499.9080457291194 141.46341463414635,499.84501495841357 141.46341463414635,499.84501495841357 C 141.46341463414635,499.84501495841357 146.34146341463415,499.9554287101044 146.34146341463415,499.9554287101044 C 146.34146341463415,499.9554287101044 151.21951219512195,499.8180852156077 151.21951219512195,499.8180852156077 C 151.21951219512195,499.8180852156077 156.09756097560975,499.9153394559902 156.09756097560975,499.9153394559902 C 156.09756097560975,499.9153394559902 160.97560975609755,499.94833481825464 160.97560975609755,499.94833481825464 C 160.97560975609755,499.94833481825464 165.85365853658536,499.71747720691536 165.85365853658536,499.71747720691536 C 165.85365853658536,499.71747720691536 170.73170731707316,499.7170734266381 170.73170731707316,499.7170734266381 C 170.73170731707316,499.7170734266381 175.60975609756096,499.6856680015181 175.60975609756096,499.6856680015181 C 175.60975609756096,499.6856680015181 180.48780487804876,499.685204884963 180.48780487804876,499.685204884963 C 180.48780487804876,499.685204884963 185.3658536585366,499.924554630154 185.3658536585366,499.924554630154 C 185.3658536585366,499.924554630154 190.2439024390244,499.75343577166694 190.2439024390244,499.75343577166694 C 190.2439024390244,499.75343577166694 195.1219512195122,499.76658027401595 195.1219512195122,499.76658027401595 C 195.1219512195122,499.76658027401595 200,499.4926288106007 200,499.4926288106007 C 200,499.4926288106007 204.8780487804878,499.70323248936205 204.8780487804878,499.70323248936205 C 204.8780487804878,499.70323248936205 209.7560975609756,499.59300867028867 209.7560975609756,499.59300867028867 C 209.7560975609756,499.59300867028867 214.6341463414634,499.89884671762854 214.6341463414634,499.89884671762854 C 214.6341463414634,499.89884671762854 219.5121951219512,499.56831798459183 219.5121951219512,499.56831798459183 C 219.5121951219512,499.56831798459183 224.390243902439,499.3470924677361 224.390243902439,499.3470924677361 C 224.390243902439,499.3470924677361 229.2682926829268,499.16418631554126 229.2682926829268,499.16418631554126 C 229.2682926829268,499.16418631554126 234.14634146341461,499.2104788322566 234.14634146341461,499.2104788322566 C 234.14634146341461,499.2104788322566 239.02439024390245,499.8040765907036 239.02439024390245,499.8040765907036 C 239.02439024390245,499.8040765907036 243.90243902439025,499.17517607371195 243.90243902439025,499.17517607371195 C 243.90243902439025,499.17517607371195 248.78048780487805,499.66901528001154 248.78048780487805,499.66901528001154 C 248.78048780487805,499.66901528001154 253.65853658536585,499.375757418371 253.65853658536585,499.375757418371 C 253.65853658536585,499.375757418371 258.5365853658536,498.74430416620714 258.5365853658536,498.74430416620714 C 258.5365853658536,498.74430416620714 263.4146341463415,499.2039467733315 263.4146341463415,499.2039467733315 C 263.4146341463415,499.2039467733315 268.2926829268293,499.2822338275746 268.2926829268293,499.2822338275746 C 268.2926829268293,499.2822338275746 273.1707317073171,499.3080112645127 273.1707317073171,499.3080112645127 C 273.1707317073171,499.3080112645127 278.0487804878049,498.8614792528015 278.0487804878049,498.8614792528015 C 278.0487804878049,498.8614792528015 282.9268292682927,499.31890602707665 282.9268292682927,499.31890602707665 C 282.9268292682927,499.31890602707665 287.8048780487805,499.0654180559883 287.8048780487805,499.0654180559883 C 287.8048780487805,499.0654180559883 292.6829268292683,498.73923680062745 292.6829268292683,498.73923680062745 C 292.6829268292683,498.73923680062745 297.5609756097561,499.99770239880837 297.5609756097561,499.99770239880837 C 297.5609756097561,499.99770239880837 302.4390243902439,499.18370868340645 302.4390243902439,499.18370868340645 C 302.4390243902439,499.18370868340645 307.3170731707317,499.9013511457095 307.3170731707317,499.9013511457095 C 307.3170731707317,499.9013511457095 312.1951219512195,499.0858199645777 312.1951219512195,499.0858199645777 C 312.1951219512195,499.0858199645777 317.0731707317073,499.06368074354396 317.0731707317073,499.06368074354396 C 317.0731707317073,499.06368074354396 321.9512195121951,498.99169685322136 321.9512195121951,498.99169685322136 C 321.9512195121951,498.99169685322136 326.8292682926829,499.7900511455965 326.8292682926829,499.7900511455965 C 326.8292682926829,499.7900511455965 331.7073170731707,498.0341769841697 331.7073170731707,498.0341769841697 C 331.7073170731707,498.0341769841697 336.5853658536585,498.4577777088165 336.5853658536585,498.4577777088165 C 336.5853658536585,498.4577777088165 341.4634146341463,498.4776869826434 341.4634146341463,498.4776869826434 C 341.4634146341463,498.4776869826434 346.3414634146341,498.14521215454977 346.3414634146341,498.14521215454977 C 346.3414634146341,498.14521215454977 351.2195121951219,497.7417897569322 351.2195121951219,497.7417897569322 C 351.2195121951219,497.7417897569322 356.0975609756097,496.3743028764938 356.0975609756097,496.3743028764938 C 356.0975609756097,496.3743028764938 360.9756097560975,499.6456567416484 360.9756097560975,499.6456567416484 C 360.9756097560975,499.6456567416484 365.8536585365854,496.95618287313204 365.8536585365854,496.95618287313204 C 365.8536585365854,496.95618287313204 370.7317073170732,497.00982786475356 370.7317073170732,497.00982786475356 C 370.7317073170732,497.00982786475356 375.609756097561,497.92776399517953 375.609756097561,497.92776399517953 C 375.609756097561,497.92776399517953 380.4878048780488,495.5160828237326 380.4878048780488,495.5160828237326 C 380.4878048780488,495.5160828237326 385.3658536585366,498.6601097355852 385.3658536585366,498.6601097355852 C 385.3658536585366,498.6601097355852 390.2439024390244,499.748464778396 390.2439024390244,499.748464778396 C 390.2439024390244,499.748464778396 395.1219512195122,495.2962181896429 395.1219512195122,495.2962181896429 C 395.1219512195122,495.2962181896429 400,498.9607285429721 400,498.9607285429721 C 400,498.9607285429721 404.8780487804878,498.7220117745617 404.8780487804878,498.7220117745617 C 404.8780487804878,498.7220117745617 409.7560975609756,498.09977604612914 409.7560975609756,498.09977604612914 C 409.7560975609756,498.09977604612914 414.6341463414634,499.43071458691867 414.6341463414634,499.43071458691867 C 414.6341463414634,499.43071458691867 419.5121951219512,497.7806696754445 419.5121951219512,497.7806696754445 C 419.5121951219512,497.7806696754445 424.390243902439,493.28987841837574 424.390243902439,493.28987841837574 C 424.390243902439,493.28987841837574 429.2682926829268,494.50419349945076 429.2682926829268,494.50419349945076 C 429.2682926829268,494.50419349945076 434.1463414634146,492.87932735861165 434.1463414634146,492.87932735861165 C 434.1463414634146,492.87932735861165 439.0243902439024,498.6915188778363 439.0243902439024,498.6915188778363 C 439.0243902439024,498.6915188778363 443.9024390243902,498.10194913682335 443.9024390243902,498.10194913682335 C 443.9024390243902,498.10194913682335 448.780487804878,498.1148392892645 448.780487804878,498.1148392892645 C 448.780487804878,498.1148392892645 453.6585365853658,491.37298906777175 453.6585365853658,491.37298906777175 C 453.6585365853658,491.37298906777175 458.5365853658536,498.38378584480256 458.5365853658536,498.38378584480256 C 458.5365853658536,498.38378584480256 463.4146341463414,490.9599176664429 463.4146341463414,490.9599176664429 C 463.4146341463414,490.9599176664429 468.29268292682923,497.0303860766468 468.29268292682923,497.0303860766468 C 468.29268292682923,497.0303860766468 473.1707317073171,495.8414585686672 473.1707317073171,495.8414585686672 C 473.1707317073171,495.8414585686672 478.0487804878049,497.3599884772825 478.0487804878049,497.3599884772825 C 478.0487804878049,497.3599884772825 482.9268292682927,496.36298561490514 482.9268292682927,496.36298561490514 C 482.9268292682927,496.36298561490514 487.8048780487805,494.44655664323653 487.8048780487805,494.44655664323653 C 487.8048780487805,494.44655664323653 492.6829268292683,488.88054581332267 492.6829268292683,488.88054581332267 C 492.6829268292683,488.88054581332267 497.5609756097561,491.2695334620059 497.5609756097561,491.2695334620059 C 497.5609756097561,491.2695334620059 502.4390243902439,493.425238239918 502.4390243902439,493.425238239918 C 502.4390243902439,493.425238239918 507.3170731707317,495.87710023778203 507.3170731707317,495.87710023778203 C 507.3170731707317,495.87710023778203 512.1951219512194,487.73846987963793 512.1951219512194,487.73846987963793 C 512.1951219512194,487.73846987963793 517.0731707317073,499.13728504267885 517.0731707317073,499.13728504267885 C 517.0731707317073,499.13728504267885 521.9512195121952,488.4499494621698 521.9512195121952,488.4499494621698 C 521.9512195121952,488.4499494621698 526.829268292683,496.2089558554846 526.829268292683,496.2089558554846 C 526.829268292683,496.2089558554846 531.7073170731708,498.97733548479164 531.7073170731708,498.97733548479164 C 531.7073170731708,498.97733548479164 536.5853658536586,487.618079369193 536.5853658536586,487.618079369193 C 536.5853658536586,487.618079369193 541.4634146341464,487.61979490171296 541.4634146341464,487.61979490171296 C 541.4634146341464,487.61979490171296 546.3414634146342,489.09403740462096 546.3414634146342,489.09403740462096 C 546.3414634146342,489.09403740462096 551.219512195122,492.14060460630145 551.219512195122,492.14060460630145 C 551.219512195122,492.14060460630145 556.0975609756098,489.57447377748844 556.0975609756098,489.57447377748844 C 556.0975609756098,489.57447377748844 560.9756097560976,497.03784510907695 560.9756097560976,497.03784510907695 C 560.9756097560976,497.03784510907695 565.8536585365854,487.0450353453407 565.8536585365854,487.0450353453407 C 565.8536585365854,487.0450353453407 570.7317073170732,495.6478015490989 570.7317073170732,495.6478015490989 C 570.7317073170732,495.6478015490989 575.609756097561,489.2071952464976 575.609756097561,489.2071952464976 C 575.609756097561,489.2071952464976 580.4878048780488,494.998133806539 580.4878048780488,494.998133806539 C 580.4878048780488,494.998133806539 585.3658536585366,491.11132306402413 585.3658536585366,491.11132306402413 C 585.3658536585366,491.11132306402413 590.2439024390244,493.9984113073628 590.2439024390244,493.9984113073628 C 590.2439024390244,493.9984113073628 595.1219512195122,493.86074476435454 595.1219512195122,493.86074476435454 C 595.1219512195122,493.86074476435454 600,481.10169807799264 600,481.10169807799264 C 600,481.10169807799264 604.8780487804878,477.69275289970835 604.8780487804878,477.69275289970835 C 604.8780487804878,477.69275289970835 609.7560975609756,483.36032240446684 609.7560975609756,483.36032240446684 C 609.7560975609756,483.36032240446684 614.6341463414634,498.52155440595664 614.6341463414634,498.52155440595664 C 614.6341463414634,498.52155440595664 619.5121951219512,493.0089193171211 619.5121951219512,493.0089193171211 C 619.5121951219512,493.0089193171211 624.390243902439,493.3279177466967 624.390243902439,493.3279177466967 C 624.390243902439,493.3279177466967 629.2682926829268,481.5084948825125 629.2682926829268,481.5084948825125 C 629.2682926829268,481.5084948825125 634.1463414634146,475.62608200138186 634.1463414634146,475.62608200138186 C 634.1463414634146,475.62608200138186 639.0243902439024,479.84152299056126 639.0243902439024,479.84152299056126 C 639.0243902439024,479.84152299056126 643.9024390243902,488.38952557154755 643.9024390243902,488.38952557154755 C 643.9024390243902,488.38952557154755 648.780487804878,473.03388205857647 648.780487804878,473.03388205857647 C 648.780487804878,473.03388205857647 653.6585365853658,475.9821958431486 653.6585365853658,475.9821958431486 C 653.6585365853658,475.9821958431486 658.5365853658536,496.34733372524744 658.5365853658536,496.34733372524744 C 658.5365853658536,496.34733372524744 663.4146341463414,480.12001934929145 663.4146341463414,480.12001934929145 C 663.4146341463414,480.12001934929145 668.2926829268292,481.2182065601716 668.2926829268292,481.2182065601716 C 668.2926829268292,481.2182065601716 673.170731707317,469.03509705770347 673.170731707317,469.03509705770347 C 673.170731707317,469.03509705770347 678.0487804878048,474.14251179191905 678.0487804878048,474.14251179191905 C 678.0487804878048,474.14251179191905 682.9268292682926,477.9257423612485 682.9268292682926,477.9257423612485 C 682.9268292682926,477.9257423612485 687.8048780487804,497.7899561395063 687.8048780487804,497.7899561395063 C 687.8048780487804,497.7899561395063 692.6829268292682,487.9301277267934 692.6829268292682,487.9301277267934 C 692.6829268292682,487.9301277267934 697.560975609756,464.5619018460006 697.560975609756,464.5619018460006 C 697.560975609756,464.5619018460006 702.4390243902438,470.3621621853033 702.4390243902438,470.3621621853033 C 702.4390243902438,470.3621621853033 707.3170731707316,489.7679000145479 707.3170731707316,489.7679000145479 C 707.3170731707316,489.7679000145479 712.1951219512194,465.6254005090687 712.1951219512194,465.6254005090687 C 712.1951219512194,465.6254005090687 717.0731707317073,460.1203494433305 717.0731707317073,460.1203494433305 C 717.0731707317073,460.1203494433305 721.951219512195,456.92228351442196 721.951219512195,456.92228351442196 C 721.951219512195,456.92228351442196 726.8292682926829,464.0348513067689 726.8292682926829,464.0348513067689 C 726.8292682926829,464.0348513067689 731.7073170731708,490.9784951987187 731.7073170731708,490.9784951987187 C 731.7073170731708,490.9784951987187 736.5853658536586,456.5709558724289 736.5853658536586,456.5709558724289 C 736.5853658536586,456.5709558724289 741.4634146341464,457.3796143652458 741.4634146341464,457.3796143652458 C 741.4634146341464,457.3796143652458 746.3414634146342,456.80397037291414 746.3414634146342,456.80397037291414 C 746.3414634146342,456.80397037291414 751.219512195122,467.3187053120997 751.219512195122,467.3187053120997 C 751.219512195122,467.3187053120997 756.0975609756098,477.64488106582087 756.0975609756098,477.64488106582087 C 756.0975609756098,477.64488106582087 760.9756097560976,466.9385989962091 760.9756097560976,466.9385989962091 C 760.9756097560976,466.9385989962091 765.8536585365854,451.84278120927047 765.8536585365854,451.84278120927047 C 765.8536585365854,451.84278120927047 770.7317073170732,461.4440788863084 770.7317073170732,461.4440788863084 C 770.7317073170732,461.4440788863084 775.609756097561,464.30595399115776 775.609756097561,464.30595399115776 C 775.609756097561,464.30595399115776 780.4878048780488,443.26627107308843 780.4878048780488,443.26627107308843 C 780.4878048780488,443.26627107308843 785.3658536585366,434.5579482135822 785.3658536585366,434.5579482135822 C 785.3658536585366,434.5579482135822 790.2439024390244,467.5074829744035 790.2439024390244,467.5074829744035 C 790.2439024390244,467.5074829744035 795.1219512195122,450.19865598145964 795.1219512195122,450.19865598145964 C 795.1219512195122,450.19865598145964 800,433.2313079470399 800,433.2313079470399 C 800,433.2313079470399 804.8780487804878,448.598255407735 804.8780487804878,448.598255407735 C 804.8780487804878,448.598255407735 809.7560975609756,421.82728735039757 809.7560975609756,421.82728735039757 C 809.7560975609756,421.82728735039757 814.6341463414634,444.10138084772024 814.6341463414634,444.10138084772024 C 814.6341463414634,444.10138084772024 819.5121951219512,451.97870652129455 819.5121951219512,451.97870652129455 C 819.5121951219512,451.97870652129455 824.390243902439,449.9972699170539 824.390243902439,449.9972699170539 C 824.390243902439,449.9972699170539 829.2682926829268,410.09477046376986 829.2682926829268,410.09477046376986 C 829.2682926829268,410.09477046376986 834.1463414634146,429.48871429367523 834.1463414634146,429.48871429367523 C 834.1463414634146,429.48871429367523 839.0243902439024,432.22514345900515 839.0243902439024,432.22514345900515 C 839.0243902439024,432.22514345900515 843.9024390243902,428.0562117597184 843.9024390243902,428.0562117597184 C 843.9024390243902,428.0562117597184 848.780487804878,389.74229298208843 848.780487804878,389.74229298208843 C 848.780487804878,389.74229298208843 853.6585365853658,437.50048009116165 853.6585365853658,437.50048009116165 C 853.6585365853658,437.50048009116165 858.5365853658536,403.85475631933116 858.5365853658536,403.85475631933116 C 858.5365853658536,403.85475631933116 863.4146341463414,395.551797676389 863.4146341463414,395.551797676389 C 863.4146341463414,395.551797676389 868.2926829268292,369.7432273922798 868.2926829268292,369.7432273922798 C 868.2926829268292,369.7432273922798 873.170731707317,384.2032877341777 873.170731707317,384.2032877341777 C 873.170731707317,384.2032877341777 878.0487804878048,403.50423376307435 878.0487804878048,403.50423376307435 C 878.0487804878048,403.50423376307435 882.9268292682926,361.0049249699402 882.9268292682926,361.0049249699402 C 882.9268292682926,361.0049249699402 887.8048780487804,350.15668673463193 887.8048780487804,350.15668673463193 C 887.8048780487804,350.15668673463193 892.6829268292682,358.0062111624337 892.6829268292682,358.0062111624337 C 892.6829268292682,358.0062111624337 897.560975609756,333.0138717631859 897.560975609756,333.0138717631859 C 897.560975609756,333.0138717631859 902.4390243902438,350.70142634168656 902.4390243902438,350.70142634168656 C 902.4390243902438,350.70142634168656 907.3170731707316,348.6795149633591 907.3170731707316,348.6795149633591 C 907.3170731707316,348.6795149633591 912.1951219512194,302.4633869318664 912.1951219512194,302.4633869318664 C 912.1951219512194,302.4633869318664 917.0731707317073,258.9045868612644 917.0731707317073,258.9045868612644 C 917.0731707317073,258.9045868612644 921.951219512195,295.6194339366724 921.951219512195,295.6194339366724 C 921.951219512195,295.6194339366724 926.8292682926829,248.99346745137777 926.8292682926829,248.99346745137777 C 926.8292682926829,248.99346745137777 931.7073170731707,204.73863099983237 931.7073170731707,204.73863099983237 C 931.7073170731707,204.73863099983237 936.5853658536585,170.72247507418405 936.5853658536585,170.72247507418405 C 936.5853658536585,170.72247507418405 941.4634146341464,202.11801321761766 941.4634146341464,202.11801321761766 C 941.4634146341464,202.11801321761766 946.3414634146342,118.39287368567165 946.3414634146342,118.39287368567165 C 946.3414634146342,118.39287368567165 951.219512195122,171.569985184289 951.219512195122,171.569985184289 C 951.219512195122,171.569985184289 956.0975609756098,121.00619811075836 956.0975609756098,121.00619811075836 C 956.0975609756098,121.00619811075836 960.9756097560976,98.91212364408796 960.9756097560976,98.91212364408796 C 960.9756097560976,98.91212364408796 965.8536585365854,19.464634386687692 965.8536585365854,19.464634386687692 C 965.8536585365854,19.464634386687692 970.7317073170732,0 970.7317073170732,0 C 970.7317073170732,0 975.609756097561,0 975.609756097561,0 C 975.609756097561,0 980.4878048780488,0 980.4878048780488,0 C 980.4878048780488,0 985.3658536585366,0 985.3658536585366,0 C 985.3658536585366,0 990.2439024390244,0 990.2439024390244,0 C 990.2439024390244,0 995.1219512195122,0 995.1219512195122,0"
+      fill="none"
+      stroke="#fbbf24"
+      stroke-width="1px"
+    />
+  </svg>
+);
+
+export default GrowthSvg;
