+++ components/ValueProp.js
@@ -1,35 +1,15 @@
 import { Parallax } from "react-scroll-parallax";
-
-// SVG exported from https://www.magicpattern.design/tools/svg-chart-generator,
-// data generated with `[1.1**i + random.randrange(0, i**2+1)) for i in range(100)]`
-// (parameters are super arbitrary)
-const GrowthSvg = () => (
-  <svg
-    width="135%" // Additional width for parallax effect
-    height="15rem" // Set height less than container, since I can't figure out how to fit it to the container height
-    preserveAspectRatio="none"
-    viewBox="0 0 990 500"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M 0,499.9651745818407 C 0,499.9651745818407 10,499.94427933094516 10,499.94427933094516 C 10,499.94427933094516 20,499.923035825868 20,499.923035825868 C 20,499.923035825868 30,499.9188219502707 30,499.9188219502707 C 30,499.9188219502707 40,499.9141866871137 40,499.9141866871137 C 40,499.9141866871137 50,499.92650060672065 50,499.92650060672065 C 50,499.92650060672065 60,499.7293521384246 60,499.7293521384246 C 60,499.7293521384246 70,499.37492842156973 70,499.37492842156973 C 70,499.37492842156973 80,499.0895385875073 80,499.0895385875073 C 80,499.0895385875073 90,499.74375639486675 90,499.74375639486675 C 90,499.74375639486675 100,499.8051955797517 100,499.8051955797517 C 100,499.8051955797517 110,499.55238483605956 110,499.55238483605956 C 110,499.55238483605956 120,497.8708286661791 120,497.8708286661791 C 120,497.8708286661791 130,498.8001852484216 130,498.8001852484216 C 130,498.8001852484216 140,497.0817170797525 140,497.0817170797525 C 140,497.0817170797525 150,497.3470954782764 150,497.3470954782764 C 150,497.3470954782764 160,497.38478616408975 160,497.38478616408975 C 160,497.38478616408975 170,498.41354652330034 170,498.41354652330034 C 170,498.41354652330034 180,495.27906919391927 180,495.27906919391927 C 180,495.27906919391927 190,498.27210522016026 190,498.27210522016026 C 190,498.27210522016026 200,496.7707260393995 200,496.7707260393995 C 200,496.7707260393995 210,493.9960892049256 210,493.9960892049256 C 210,493.9960892049256 220,496.7215255596301 220,496.7215255596301 C 220,496.7215255596301 230,491.81761816946283 230,491.81761816946283 C 230,491.81761816946283 240,496.5575167246313 240,496.5575167246313 C 240,496.5575167246313 250,498.2122473994374 250,498.2122473994374 C 250,498.2122473994374 260,498.5750073917581 260,498.5750073917581 C 260,498.5750073917581 270,492.177863029526 270,492.177863029526 C 270,492.177863029526 280,495.4580343607594 280,495.4580343607594 C 280,495.4580343607594 290,493.0570969217312 290,493.0570969217312 C 290,493.0570969217312 300,491.0516296202072 300,491.0516296202072 C 300,491.0516296202072 310,497.7818178882038 310,497.7818178882038 C 310,497.7818178882038 320,489.1827453388077 320,489.1827453388077 C 320,489.1827453388077 330,489.4574699099927 330,489.4574699099927 C 330,489.4574699099927 340,492.3715733002428 340,492.3715733002428 C 340,492.3715733002428 350,489.16572754639304 350,489.16572754639304 C 350,489.16572754639304 360,481.17990242186164 360,481.17990242186164 C 360,481.17990242186164 370,486.41794940671406 370,486.41794940671406 C 370,486.41794940671406 380,475.29469709552137 380,475.29469709552137 C 380,475.29469709552137 390,496.73878145505586 390,496.73878145505586 C 390,496.73878145505586 400,493.61791979327865 400,493.61791979327865 C 400,493.61791979327865 410,492.3110637439482 410,492.3110637439482 C 410,492.3110637439482 420,477.31946934209026 420,477.31946934209026 C 420,477.31946934209026 430,481.4471043212523 430,481.4471043212523 C 430,481.4471043212523 440,474.150343165989 440,474.150343165989 C 440,474.150343165989 450,470.66339915226234 450,470.66339915226234 C 450,470.66339915226234 460,476.4691777545616 460,476.4691777545616 C 460,476.4691777545616 470,460.41403475524635 470,460.41403475524635 C 470,460.41403475524635 480,457.3905007083479 480,457.3905007083479 C 480,457.3905007083479 490,484.1642221718311 490,484.1642221718311 C 490,484.1642221718311 500,468.2952578600748 500,468.2952578600748 C 500,468.2952578600748 510,470.95107610413095 510,470.95107610413095 C 510,470.95107610413095 520,483.578320213586 520,483.578320213586 C 520,483.578320213586 530,494.33225982874274 530,494.33225982874274 C 530,494.33225982874274 540,467.56458245947755 540,467.56458245947755 C 540,467.56458245947755 550,456.08308803984596 550,456.08308803984596 C 550,456.08308803984596 560,482.65815867106846 560,482.65815867106846 C 560,482.65815867106846 570,466.680378511027 570,466.680378511027 C 570,466.680378511027 580,483.9755115378753 580,483.9755115378753 C 580,483.9755115378753 590,476.18632715566554 590,476.18632715566554 C 590,476.18632715566554 600,464.1478714156618 600,464.1478714156618 C 600,464.1478714156618 610,476.04255692903104 610,476.04255692903104 C 610,476.04255692903104 620,452.9361364426062 620,452.9361364426062 C 620,452.9361364426062 630,479.77461385554886 630,479.77461385554886 C 630,479.77461385554886 640,448.4830525491311 640,448.4830525491311 C 640,448.4830525491311 650,439.8435921753917 650,439.8435921753917 C 650,439.8435921753917 660,421.8375599206884 660,421.8375599206884 C 660,421.8375599206884 670,464.6574739163603 670,464.6574739163603 C 670,464.6574739163603 680,442.6361480781389 680,442.6361480781389 C 680,442.6361480781389 690,447.8332029170609 690,447.8332029170609 C 690,447.8332029170609 700,453.23827574734946 700,453.23827574734946 C 700,453.23827574734946 710,462.3982419567691 710,462.3982419567691 C 710,462.3982419567691 720,396.6348916616517 720,396.6348916616517 C 720,396.6348916616517 730,444.1869321630914 730,444.1869321630914 C 730,444.1869321630914 740,402.81031932437753 740,402.81031932437753 C 740,402.81031932437753 750,388.10957518912926 750,388.10957518912926 C 750,388.10957518912926 760,378.4737463579464 760,378.4737463579464 C 760,378.4737463579464 770,372.1038816790906 770,372.1038816790906 C 770,372.1038816790906 780,401.86570702492435 780,401.86570702492435 C 780,401.86570702492435 790,384.80859075028513 790,384.80859075028513 C 790,384.80859075028513 800,415.41253435987903 800,415.41253435987903 C 800,415.41253435987903 810,349.92890683094265 810,349.92890683094265 C 810,349.92890683094265 820,389.72307552949076 820,389.72307552949076 C 820,389.72307552949076 830,339.40534631513236 830,339.40534631513236 C 830,339.40534631513236 840,301.1098514519409 840,301.1098514519409 C 840,301.1098514519409 850,363.97299522723245 850,363.97299522723245 C 850,363.97299522723245 860,344.8051929281017 860,344.8051929281017 C 860,344.8051929281017 870,315.1727115118608 870,315.1727115118608 C 870,315.1727115118608 880,226.30955383403608 880,226.30955383403608 C 880,226.30955383403608 890,283.82050089706956 890,283.82050089706956 C 890,283.82050089706956 900,249.86564660384917 900,249.86564660384917 C 900,249.86564660384917 910,165.64725912253846 910,165.64725912253846 C 910,165.64725912253846 920,262.6732252722047 920,262.6732252722047 C 920,262.6732252722047 930,236.57590190640963 930,236.57590190640963 C 930,236.57590190640963 940,210.4859763787054 940,210.4859763787054 C 940,210.4859763787054 950,74.65362441481653 950,74.65362441481653 C 950,74.65362441481653 960,105.95583843761824 960,105.95583843761824 C 960,105.95583843761824 970,99.3482598328884 970,99.3482598328884 C 970,99.3482598328884 980,41.43854155135858 980,41.43854155135858 C 980,41.43854155135858 990,0 990,0,L 1000 500,L 0 500Z"
-      fill="#fbbf241a"
-    />
-    <path
-      d="M 0,499.9651745818407 C 0,499.9651745818407 10,499.94427933094516 10,499.94427933094516 C 10,499.94427933094516 20,499.923035825868 20,499.923035825868 C 20,499.923035825868 30,499.9188219502707 30,499.9188219502707 C 30,499.9188219502707 40,499.9141866871137 40,499.9141866871137 C 40,499.9141866871137 50,499.92650060672065 50,499.92650060672065 C 50,499.92650060672065 60,499.7293521384246 60,499.7293521384246 C 60,499.7293521384246 70,499.37492842156973 70,499.37492842156973 C 70,499.37492842156973 80,499.0895385875073 80,499.0895385875073 C 80,499.0895385875073 90,499.74375639486675 90,499.74375639486675 C 90,499.74375639486675 100,499.8051955797517 100,499.8051955797517 C 100,499.8051955797517 110,499.55238483605956 110,499.55238483605956 C 110,499.55238483605956 120,497.8708286661791 120,497.8708286661791 C 120,497.8708286661791 130,498.8001852484216 130,498.8001852484216 C 130,498.8001852484216 140,497.0817170797525 140,497.0817170797525 C 140,497.0817170797525 150,497.3470954782764 150,497.3470954782764 C 150,497.3470954782764 160,497.38478616408975 160,497.38478616408975 C 160,497.38478616408975 170,498.41354652330034 170,498.41354652330034 C 170,498.41354652330034 180,495.27906919391927 180,495.27906919391927 C 180,495.27906919391927 190,498.27210522016026 190,498.27210522016026 C 190,498.27210522016026 200,496.7707260393995 200,496.7707260393995 C 200,496.7707260393995 210,493.9960892049256 210,493.9960892049256 C 210,493.9960892049256 220,496.7215255596301 220,496.7215255596301 C 220,496.7215255596301 230,491.81761816946283 230,491.81761816946283 C 230,491.81761816946283 240,496.5575167246313 240,496.5575167246313 C 240,496.5575167246313 250,498.2122473994374 250,498.2122473994374 C 250,498.2122473994374 260,498.5750073917581 260,498.5750073917581 C 260,498.5750073917581 270,492.177863029526 270,492.177863029526 C 270,492.177863029526 280,495.4580343607594 280,495.4580343607594 C 280,495.4580343607594 290,493.0570969217312 290,493.0570969217312 C 290,493.0570969217312 300,491.0516296202072 300,491.0516296202072 C 300,491.0516296202072 310,497.7818178882038 310,497.7818178882038 C 310,497.7818178882038 320,489.1827453388077 320,489.1827453388077 C 320,489.1827453388077 330,489.4574699099927 330,489.4574699099927 C 330,489.4574699099927 340,492.3715733002428 340,492.3715733002428 C 340,492.3715733002428 350,489.16572754639304 350,489.16572754639304 C 350,489.16572754639304 360,481.17990242186164 360,481.17990242186164 C 360,481.17990242186164 370,486.41794940671406 370,486.41794940671406 C 370,486.41794940671406 380,475.29469709552137 380,475.29469709552137 C 380,475.29469709552137 390,496.73878145505586 390,496.73878145505586 C 390,496.73878145505586 400,493.61791979327865 400,493.61791979327865 C 400,493.61791979327865 410,492.3110637439482 410,492.3110637439482 C 410,492.3110637439482 420,477.31946934209026 420,477.31946934209026 C 420,477.31946934209026 430,481.4471043212523 430,481.4471043212523 C 430,481.4471043212523 440,474.150343165989 440,474.150343165989 C 440,474.150343165989 450,470.66339915226234 450,470.66339915226234 C 450,470.66339915226234 460,476.4691777545616 460,476.4691777545616 C 460,476.4691777545616 470,460.41403475524635 470,460.41403475524635 C 470,460.41403475524635 480,457.3905007083479 480,457.3905007083479 C 480,457.3905007083479 490,484.1642221718311 490,484.1642221718311 C 490,484.1642221718311 500,468.2952578600748 500,468.2952578600748 C 500,468.2952578600748 510,470.95107610413095 510,470.95107610413095 C 510,470.95107610413095 520,483.578320213586 520,483.578320213586 C 520,483.578320213586 530,494.33225982874274 530,494.33225982874274 C 530,494.33225982874274 540,467.56458245947755 540,467.56458245947755 C 540,467.56458245947755 550,456.08308803984596 550,456.08308803984596 C 550,456.08308803984596 560,482.65815867106846 560,482.65815867106846 C 560,482.65815867106846 570,466.680378511027 570,466.680378511027 C 570,466.680378511027 580,483.9755115378753 580,483.9755115378753 C 580,483.9755115378753 590,476.18632715566554 590,476.18632715566554 C 590,476.18632715566554 600,464.1478714156618 600,464.1478714156618 C 600,464.1478714156618 610,476.04255692903104 610,476.04255692903104 C 610,476.04255692903104 620,452.9361364426062 620,452.9361364426062 C 620,452.9361364426062 630,479.77461385554886 630,479.77461385554886 C 630,479.77461385554886 640,448.4830525491311 640,448.4830525491311 C 640,448.4830525491311 650,439.8435921753917 650,439.8435921753917 C 650,439.8435921753917 660,421.8375599206884 660,421.8375599206884 C 660,421.8375599206884 670,464.6574739163603 670,464.6574739163603 C 670,464.6574739163603 680,442.6361480781389 680,442.6361480781389 C 680,442.6361480781389 690,447.8332029170609 690,447.8332029170609 C 690,447.8332029170609 700,453.23827574734946 700,453.23827574734946 C 700,453.23827574734946 710,462.3982419567691 710,462.3982419567691 C 710,462.3982419567691 720,396.6348916616517 720,396.6348916616517 C 720,396.6348916616517 730,444.1869321630914 730,444.1869321630914 C 730,444.1869321630914 740,402.81031932437753 740,402.81031932437753 C 740,402.81031932437753 750,388.10957518912926 750,388.10957518912926 C 750,388.10957518912926 760,378.4737463579464 760,378.4737463579464 C 760,378.4737463579464 770,372.1038816790906 770,372.1038816790906 C 770,372.1038816790906 780,401.86570702492435 780,401.86570702492435 C 780,401.86570702492435 790,384.80859075028513 790,384.80859075028513 C 790,384.80859075028513 800,415.41253435987903 800,415.41253435987903 C 800,415.41253435987903 810,349.92890683094265 810,349.92890683094265 C 810,349.92890683094265 820,389.72307552949076 820,389.72307552949076 C 820,389.72307552949076 830,339.40534631513236 830,339.40534631513236 C 830,339.40534631513236 840,301.1098514519409 840,301.1098514519409 C 840,301.1098514519409 850,363.97299522723245 850,363.97299522723245 C 850,363.97299522723245 860,344.8051929281017 860,344.8051929281017 C 860,344.8051929281017 870,315.1727115118608 870,315.1727115118608 C 870,315.1727115118608 880,226.30955383403608 880,226.30955383403608 C 880,226.30955383403608 890,283.82050089706956 890,283.82050089706956 C 890,283.82050089706956 900,249.86564660384917 900,249.86564660384917 C 900,249.86564660384917 910,165.64725912253846 910,165.64725912253846 C 910,165.64725912253846 920,262.6732252722047 920,262.6732252722047 C 920,262.6732252722047 930,236.57590190640963 930,236.57590190640963 C 930,236.57590190640963 940,210.4859763787054 940,210.4859763787054 C 940,210.4859763787054 950,74.65362441481653 950,74.65362441481653 C 950,74.65362441481653 960,105.95583843761824 960,105.95583843761824 C 960,105.95583843761824 970,99.3482598328884 970,99.3482598328884 C 970,99.3482598328884 980,41.43854155135858 980,41.43854155135858 C 980,41.43854155135858 990,0 990,0"
-      fill="none"
-      stroke="#fbbf24"
-      stroke-width="1px"
-    />
-  </svg>
-);
+import GrowthSvg from "./GrowthSvg";
 
 const Growth = () => (
   <div className="relative pt-12 pb-6">
     {/* Not using top-0 so it clips to the bottom, since I can't figure out
         how to fit the SVG inside the container height */}
-    <div className="absolute left-0 right-0 bottom-0 overflow-hidden z-0">
-      <Parallax x={[-15, -35]}>
+    <div
+      className="absolute right-0 bottom-0 overflow-hidden z-0"
+      style={{ left: "-25%" }}
+    >
+      <Parallax x={[7.5, 0]}>
         <GrowthSvg />
       </Parallax>
     </div>
